Can choose between game and editor (see details)
Add scene trait. Game loop has a stack of scenes and will use
the one on top
Add Gui trait. Each scene has a Gui that will be passed to the
rendering system (might be empty)
Had to hack a bit for passing scene members to the rendering system
(see get_parts_mut)